refactor: move IdParser to separate file
